Set DOB and email properties in Janus user if they are specified during registration.
It is possible to create an account without an email address or DOB. This change allows these properties to be set in Janus if they are subsequently specified during the registration process. This is done when confirming the step, so the properties will not be editable during the registration process once this has happened.

* Set DOB form type input to DateTime
* Apply registration step values over the top of any set in the User entity
* Update User if the DOB or email was not set but has been provided during the registration