Add default implementation of processRequest in AbstractRegistrationStep
Should work in cases where the data only needs to be persisted to Minerva with no additional updates in other systems.

Just requires a "childFormOrder" class property to be set with the indexes being the order that is sent to Minerva.

For example,

protected $childFormOrder = array(
1 => "foo",
3 => "bar",
4 => "baz",
);

would sent the appropriate values from the "foo", "bar" and "baz" form children to Minerva with the orders set as 1, 3 and 4, respectively.