Update AbstractRegistrationStep::processRequest to not require a Request to be passed
Allows a form to be bound in an overridden processRequest method but still use the parent processRequest.

For example, I can bind the form to allow access to the data to communicate with external APIs but also use the form data to persist to Minerva, as per the default action.

* Update method signature and DocBlock
* Update child implementations of processRequest