Weekend accommodation step for residential courses
The form elements for this step are built up dynamically based on the options that are chosen. The end result is a set of fields that are appropriate for the options the person has chosen.

Using the PRE_BIND event, unfortunately, means that the constraints are validated when an option is chosen, before the customer has had a chance to select a valid option.